% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_adjust_expo.R
\name{adjust_exposition}
\alias{adjust_exposition}
\title{Adjust Exposition for Static Matching}
\usage{
adjust_exposition(
  matched_cohort,
  outcome_date_col,
  censoring_date_col,
  immunization_date,
  start_cohort,
  end_cohort
)
}
\arguments{
\item{matched_cohort}{\code{data.frame} with matched cohort from \code{match_cohort_}.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date.}

\item{start_cohort}{Start date of the study.}

\item{end_cohort}{End date of the study.}
}
\value{
\code{data.frame} with matched population and corrected exposition times.
}
\description{
This function removes the pairs whose exposition times
do not match. This happens when the outcome of the unvaccinated individual
occurs before the vaccination date of their partner.
}
\keyword{internal}
