% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_warnings.R
\name{capture_warnings}
\alias{capture_warnings}
\title{Capture Warnings}
\usage{
capture_warnings(expr, warn_env = new.env())
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{warn_env}{An optional environment to store the warnings.
If not provided, a new environment is created.}
}
\value{
A list containing two elements:
\item{result}{The result of evaluating the expression.}
\item{warnings}{A character vector of the warnings generated during
the evaluation of the expression.}
}
\description{
This function captures the warnings generated by an
expression without printing them to the console. It returns both the
result of the expression and the captured warnings. It is useful for
avoiding warnings appearing in the middle of a routine (for instance,
an iterative routine).
}
\keyword{internal}
