% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_validations.R
\name{check_dataset}
\alias{check_dataset}
\title{Check Correctness of \code{data_set}}
\usage{
check_dataset(data_set, columns)
}
\arguments{
\item{data_set}{\code{data.frame}.}

\item{columns}{A vector of column names that should be present in
\code{data_set}.}
}
\value{
This function does not return a value.
}
\description{
This auxiliary function checks that the type of the object
provided in \code{data_set} is a \code{data.frame} and that the specified columns
are included in it.
}
\keyword{internal}
