% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_data_wrangling.R
\name{check_minimum_dataset}
\alias{check_minimum_dataset}
\title{Checks data before calculating vaccine effectiveness using the screening
method \code{\link[=st_effectiveness]{st_effectiveness()}}}
\usage{
check_minimum_dataset(
  data,
  screening_date,
  lower_age_limit,
  upper_age_limit,
  vaccinated_population,
  unvaccinated_population,
  vaccinated_outcome,
  unvaccinated_outcome
)
}
\arguments{
\item{data}{A data frame with case, hospitalisation, and ICU information}

\item{screening_date}{Specific date the data was collected}

\item{lower_age_limit}{Lower bound of age range}

\item{upper_age_limit}{Upper bound of age range}

\item{vaccinated_population}{Character string with the name of the data column
with the vaccinated population data}

\item{unvaccinated_population}{Character string with the name of the data column
with the unvaccinated population data}

\item{vaccinated_outcome}{Character string with the name of the data column
with the outcome data for vaccinated population data}

\item{unvaccinated_outcome}{Character string with the name of the data column
with the outcome data for unvaccinated population data}
}
\value{
Nothing, used for side-effects, fails if not correct format
}
\description{
Checks data before calculating vaccine effectiveness using the screening
method \code{\link[=st_effectiveness]{st_effectiveness()}}
}
