% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{cox_model}
\alias{cox_model}
\title{Internal function to calculate Cox-PH model and related metrics.}
\usage{
cox_model(
  data_set,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status
)
}
\value{
List with data from Cox model:
hr - hazard ratio (CI95\%)
p_value
\code{{survival}} object with model
\code{{survival}} object with Schoenfeld test
}
\description{
Internal function to calculate Cox-PH model and related metrics.
}
\keyword{internal}
