% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{cox_model_prediction}
\alias{cox_model_prediction}
\title{Internal function to calculate prediction from Cox model}
\usage{
cox_model_prediction(cox_model, vaccinated_status, unvaccinated_status)
}
\arguments{
\item{cox_model}{Result from \code{cox_model} function}
}
\value{
\code{data.frame} containing
\code{time}: time-to-event until \code{at}
\code{logtime}: log of time-to-event
\code{hazard}: estimated hazard from model
\code{surv}: estimated survival probability from model
\code{loglog}: -log(-log(survival probability))
\code{strata}: vaccinated/unvaccinated status
}
\description{
Internal function to calculate prediction from Cox model
}
\keyword{internal}
