% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn_effectiveness.R
\name{estimate_tn_eff}
\alias{estimate_tn_eff}
\title{Estimate vaccine effectiveness using the test negative design}
\usage{
estimate_tn_eff(data, status_vacc_col, outcome_status_col)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{status_vacc_col}{name of the column containing the vaccination
status}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}
}
\value{
The vaccine effectiveness value
}
\description{
Estimate vaccine effectiveness using the test negative design
}
\examples{
# Load the dataset for this example
data(testnegdata)
estimate_tn_eff(testnegdata, "vaccine12mo_yn", "flu_final_posneg")
}
