% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{extract_surv_model}
\alias{extract_surv_model}
\title{Internal function to extract summary output from \code{{survival}} models.}
\usage{
extract_surv_model(model, start_cohort, end_cohort)
}
\arguments{
\item{model}{\code{{survival}} object containing the model}
}
\value{
\code{data.frame} with survival data
}
\description{
Internal function to extract summary output from \code{{survival}} models.
}
\keyword{internal}
