% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_utils.R
\name{get_censoring_after_match}
\alias{get_censoring_after_match}
\title{Create Censoring date by Pairs}
\usage{
get_censoring_after_match(data_set, outcome_date_col, censoring_date_col)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. \code{NULL} by default.}
}
\description{
This function creates the censoring date of the pairs by
inheriting the minimum date in which any of the partners has a
censoring event. Two conditions are checked to inherit a censoring date
in a pair.
\enumerate{
\item Individual censoring occurs before individual event;
\item If an outcome happens before the censoring of the partner
no censoring is inherited by the other.
}
}
\keyword{internal}
