% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_matching.R
\name{get_censoring_date_match}
\alias{get_censoring_date_match}
\title{Censor couple after matching}
\usage{
get_censoring_date_match(data, outcome_date_col, censoring_date_col)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_date_col}{name of the column that contains
the outcome dates}

\item{censoring_date_col}{name of the column that contains
the censoring date. NULL by default}
}
\description{
This function censors a couple whether the case or the control
have a censoring date. It imputes the censoring date to the whole couple
using the matching id provided in subclass. This column comes with the output
of \code{match_cohort}.
}
\examples{
# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# match cohort
matched_cohort <- match_cohort(data = cohortdata,
  status_vacc_col = "vaccine_status",
  nearest = c(age = 1),
  exact = "sex"
)

# add column with censoring date for match
matched_cohort$censoring_date_match <-  get_censoring_date_match(
  data = matched_cohort,
  outcome_date_col = "death_date",
  censoring_date_col = "death_other_causes"
)

# view data with added column
head(matched_cohort)
}
