% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match.R
\name{get_dataset.match}
\alias{get_dataset.match}
\title{Function for Extracting Matched Dataset}
\usage{
\method{get_dataset}{match}(object, ...)
}
\arguments{
\item{object}{Object of the class \code{match}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
\code{data.frame} extracted from \code{match_cohort}.
}
\description{
This function extracts the matched datasets generated
by \code{match_cohort}.
}
\keyword{internal}
