% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_dataset}
\alias{get_dataset}
\title{Generic Function for Extracting Datasets}
\usage{
get_dataset(object, ...)
}
\arguments{
\item{object}{An object from which to extract the dataset.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
The get_dataset extracted from the object.
}
\description{
This function serves as a generic method to extract datasets
from objects of various classes. It dispatches to specific methods based
on the class of the input object.
}
\keyword{internal}
