% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_time_to_event}
\alias{get_time_to_event}
\title{Function to construct the time-to-event}
\usage{
get_time_to_event(
  data,
  outcome_date_col,
  start_cohort,
  end_cohort,
  start_from_immunization = FALSE,
  immunization_date_col = FALSE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_date_col}{name of the column that contains
the outcome dates}

\item{start_cohort}{start date of the study}

\item{end_cohort}{end date of the study}

\item{start_from_immunization}{TRUE: starts counting time-to-event from
immunization date if available
FALSE: starts counting time-to-event from the start date of the cohort study}

\item{immunization_date_col}{name of the column that contains the
immunization date. Required if start_from_immunization = TRUE}
}
\value{
time-to-event
}
\description{
This function returns a column with the time-to-event in days occurred
until a reference outcome. The starting point to count the time-to-event
can be the immunization date, supossing that the vaccinate population
enters to the study, when they are vaccinated
(start_from_immunization=TRUE). Or the beginning of the study, if all the
cohort is known at this point (start_from_immunization=FALSE). In this last
case, it is not necessary to pass the argument {immunization_date_col}
}
\examples{
\dontrun{
data(cohortdata)
cohortdata$immunization_death <-
  get_immunization_date(cohortdata,
                       "death_date",
                       1,
                       1,
                       c("vaccine_date_1", "vaccine_date_2"),
                       "2021-12-31",
                       take_first = FALSE)
cohortdata$time_to_death <- get_time_to_event(cohortdata, "death_date",
                                            "2021-01-01", "2021-12-31",
                                            TRUE, "immunization_death")
}
}
