% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{km_model}
\alias{km_model}
\title{Internal function to calculate Kaplan-Meier model and related metrics.}
\usage{
km_model(
  data,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status,
  start_cohort,
  end_cohort
)
}
\arguments{
\item{data}{dataset with cohort information (see example).}

\item{outcome_status_col}{name of the column containing status of the event
(must be a binary column). Default is \code{outcome_status}.}

\item{time_to_event_col}{name of the column containing the time-to-event.
Default is \code{time_to_event}.}

\item{vacc_status_col}{name of the column containing the vaccination.
Default is \code{vaccine_status}.}

\item{vaccinated_status}{string assigned to the vaccinated population.
Default is \code{v}.}

\item{unvaccinated_status}{string assigned to the unvaccinated population.
Default is \code{u}.}

\item{start_cohort}{start date of the study.}

\item{end_cohort}{end date of the study.}
}
\value{
Data frame with data from KM model
"time", "date", "strata",
"n.risk", "n.event", "n.censor",
"surv", "lower", "upper",
"cumincidence", "cumincidence_lower", "cumincidence_upper"
}
\description{
Internal function to calculate Kaplan-Meier model and related metrics.
}
\keyword{internal}
