% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match.R, R/coh_matching.R
\name{match_cohort}
\alias{match_cohort}
\title{Match Cohort to Reduce Observational Bias}
\usage{
match_cohort(data, status_vacc_col, exact = NULL, nearest = NULL)

match_cohort(data, status_vacc_col, exact = NULL, nearest = NULL)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{status_vacc_col}{name of the column containing the information
of the vaccination status.}

\item{exact}{name(s) of column(s) for \code{exact} matching.
Default to \code{NULL}.}

\item{nearest}{named vector with name(s) of column(s) for \code{nearest}
matching and caliper(s) for each variable.
e.g. \code{nearest = c("characteristic1" = n1, "characteristic2" = n2)},
where \code{n1} and \code{n2} are the calipers. Default to \code{NULL}.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. NULL by default.}

\item{immunization_date_col}{Name of the column that contains the
immunization date to set the beginning of the follow-up period
(\code{t0_follow_up}). Default is \code{immunization_date}.}
}
\value{
\code{match} object. List with results from static match: \code{match}: data
frame with adjusted cohort, \code{summary}: matching summary, \code{balance_all}:
balance of the cohort before matching, \code{balance_matched}: balance of the
cohort after matching.

Four columns are added to the structure provided in \code{data}: \code{subclass}: ID of
matched pair, \code{t0_follow_up}: beginning of follow-up period for pair,
\code{time_to_event}: time to event, and \code{outcome_status}: outcome status
(1:positive, 0: negative).

data frame with matched population. Two columns are added
to the structure provided in \code{data}:
\code{prop_score} (propensity score of the match),
\code{subclass} (id of matched pair)
}
\description{
This function builds pairs of vaccinated and unvaccinated
individuals with similar characteristics. The function relies on the matching
algorithm implemented in the \code{{MatchIt}} package, setting, by default,
\code{method = "nearest"}, \code{ratio = 1}, and \code{distance = "mahalanobis"}.
Exact and near characteristics are accepted, passed in the parameters \code{exact}
and \code{nearest}, respectively. The parameter \code{nearest} must be provided
together with the calipers as a named vector (e.g.,
\code{nearest = c(characteristic1 = n1, characteristic2 = n2)}, where \code{n1}
and \code{n2} are the calipers). The default matching \code{method} of the function is
\code{static}. This means that pairs are matched once, without taking into
account their vaccination, censoring, and outcome dates. After this, the
pairs whose exposition time do not coincide are removed to avoid negative
time-to-events.
The function returns a matched and adjusted by exposition cohort, with
information of the beginning of follow-up period of pairs (\code{t0_follow_up}),
corresponding to the vaccination date of the vaccinated individual, the
individual time-to-event (\code{time_to_event}) and the outcome status
(\code{outcome_status}), both taking into account the right-censoring dates.
pairs are censored if the vaccinated or unvaccinated partner was previously
censored (i.e., if \code{censoring_date_col} is informed) and the censor occurs
before their outcomes. Rolling calendar matching method will be included in
future releases.

This function builds pairs of vaccinated - unvaccinated
individuals with similar characteristics. The function relies on  the
propensity score matching algorithm implemented in MatchIt package.
By default the function uses \code{method = "nearest"}, \code{ratio = 1},
\code{distance = "glm"} to match the data.
Exact and near characteristics are accepted for the matching criteria.
These are passed in the parameters \code{exact} and \code{nearest}, respectively.
Parameters \code{nearest} and \code{caliper} must be provided together. In this case,
the calipers must be passed as a named vector containing each of
the variables provided in \code{nearest}
(e.g. \verb{nearest = c("characteristic1", "characteristic2"), caliper = c(characteristic1 = n1, characteristic2 = n2)},
where \code{n1} and \code{n2} are the calipers).
\code{caliper} is ignored (set to NULL) when \code{nearest} is not provided.
}
\examples{
# Define start and end dates of the study
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")

# Create data frame with information on immunization
cohortdata <- make_immunization(
  data = cohortdata,
  outcome_date_col = "death_date",
  censoring_date_col = "death_other_causes",
  immunization_delay = 14,
  vacc_date_col = "vaccine_date_2",
  end_cohort = end_cohort
)

# Match the data
matching <- match_cohort(
  data = cohortdata,
  outcome_date_col = "death_date",
  censoring_date_col = "death_other_causes",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  method = "static",
  exact = "sex",
  nearest = c(age = 1)
)

# Check match balance and summary
summary(matching)

# Extract matched data
cohortdata_match <- dataset(matching)

# View of mached cohort
head(cohortdata_match)

# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# match cohort
matched_cohort <- match_cohort(data = cohortdata,
  status_vacc_col = "vaccine_status",
  nearest = c(age = 1),
  exact = "sex"
)

# view matched data
head(matched_cohort)
}
