% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_utils.R
\name{match_couple_info}
\alias{match_couple_info}
\title{Match couple information}
\usage{
match_couple_info(data, column_to_match, criteria = c("min", "max"))
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{column_to_match}{name of the column that contains the information
shared by the couples}

\item{criteria}{selection criteria when both individuals bring information.
This can be minimum (min) or maximum (max) value}
}
\description{
This function matches numeric/date information of a
couple provided in the column \code{to_match}. This matching procedure is
required for sharing the censoring information and the exposure times
when adjusting and removing negative exposures generated by uncoherent
exposure dates.
}
\keyword{internal}
