% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vaccineff_data.R
\name{plot.vaccineff_data}
\alias{plot.vaccineff_data}
\title{Function for Extracting Vaccineff Data Plot}
\usage{
\method{plot}{vaccineff_data}(x, date_interval = NULL, cumulative = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{vaccineff_data}.}

\item{date_interval}{If NULL, the function calculates the coverage interval}

\item{cumulative}{If \code{TRUE}, returns the cumulative number of doses over the
time window.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Plot extracted from \code{vaccineff}.
}
\description{
This function returns a plot of the vaccine coverage or the
cumulative coverage (if cumulative = TRUE). The return is a 2-axis \code{ggplot2}
element with the number of vaccines per date on the left axis and the
coverage per date on the right axis. When a matching routine is performed,
the left axis also accounts for the doses of the matched cohort.
}
