% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Function to plot the vaccine coverage in a cohort}
\usage{
plot_coverage(
  data,
  vacc_date_col,
  unit = c("day", "month", "year"),
  doses_count_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = FALSE,
  cumulative = FALSE
)
}
\arguments{
\item{doses_count_color}{color assigned to the doses count}

\item{coverage_color}{color assigned to the coverage calculation}

\item{inheritParams}{coh_coverage}
}
\value{
2-axis ggplot2 plot of vaccine coverage and daily doses
}
\description{
The function relies on the implementation of the function coh_coverage()
It returns a plot of the vaccine coverage or the cumulative coverage (if
cumulative = TRUE).
The return is a 2-axis ggplot2 element with the number of vaccines per date
in the left-axes and the coverage per date in the right-axes.
It is possible to manipulate the colors, labels, legend and most of
the graphic elemnts.To do so follow the convention:
c("c1" = "color1", "c2" = "color2")
}
\examples{
data("cohortdata")
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)
plot_coverage(
  data = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  daily_doses_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = date_interval,
  cumulative = FALSE
)
}
