% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_plot.R
\name{plot_loglog}
\alias{plot_loglog}
\title{Internal function to plot loglog test.}
\usage{
plot_loglog(km, vaccinated_status, unvaccinated_status)
}
\arguments{
\item{vaccinated_status}{string assigned to the vaccinated population}

\item{unvaccinated_status}{string assigned to the unvaccinated population}

\item{surv_plot}{survival plot generated by \code{plot_survival}}
}
\value{
loglog plot
}
\description{
Internal function to plot loglog test.
}
\keyword{internal}
