% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_vaccineff.R
\name{print.summary.vaccineff}
\alias{print.summary.vaccineff}
\title{Print summary of VE Results}
\usage{
\method{print}{summary.vaccineff}(x, ...)
}
\arguments{
\item{x}{Object of the class \code{summary.vaccineff}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Summary of the results from \code{estimate_vaccineff}.
}
\description{
Print summary of \code{vaccineff} object.
}
