% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vaccineff_data.R
\name{print.summary_vaccineff_data}
\alias{print.summary_vaccineff_data}
\title{Print Summary of Vaccineff Data}
\usage{
\method{print}{summary_vaccineff_data}(x, ...)
}
\arguments{
\item{x}{Object of the class \code{summary.vaccineff_data}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Summary of the results from vaccineff data
}
\description{
Summarizes the results of \code{make_vaccineff_data}.
}
