% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stuff.R
\name{read_stuff}
\alias{read_stuff}
\title{Example of a function to read stuff}
\usage{
read_stuff(x)
}
\arguments{
\item{x}{the path to the file to read stuff from}
}
\description{
This function reads stuff into R. More details...
}
\examples{
some_path <- "path to your file"
read_stuff(some_path)
}
\author{
Thibaut Jombart
}
