% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningdata.R
\docType{data}
\name{screeningdata}
\alias{screeningdata}
\title{Screening method data on vaccineff}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
screeningdata
}
\description{
Data from
}
\examples{
screeningdata
}
\keyword{datasets}
