% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn_data_wrangling.R
\name{set_binary_test}
\alias{set_binary_test}
\title{Function to assign the outcome status for the result of a test}
\usage{
set_binary_test(data, col_name, neg_cond, status = c(TRUE, FALSE))
}
\arguments{
\item{data}{dataset with at least one column to generate the binary status}

\item{col_name}{name of the column containing the information for the
status}

\item{neg_cond}{condition value in \code{data$colname} storing negative
test results}

\item{status_col}{binary set of status, e.g. c(1, 0). The first element of
the vector must be the status when the condition is satisfied.}
}
\value{
status
}
\description{
Function to assign the outcome status for the result of a test
}
\examples{
data("testnegdata")
testnegdata$test_binary <- set_status(
data = testnegdata,
col_name = "pcr_result",
status = c(TRUE, FALSE))

testnegdata$pcr_result_bin <- set_binary_test(testnegdata,
                                              "pcr_result",
                                              "Neg",
                                              c(FALSE, TRUE))
testnegdata$rapid_test_bin <- set_binary_test(testnegdata,
                                              "rapid_result",
                                              "Negative",
                                              c(FALSE, TRUE))
testnegdata$vacc_status_bin <- set_binary_test(testnegdata,
                                              "vacc_status",
                                              "No",
                                              c(FALSE, TRUE))
}
