% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{set_event_status}
\alias{set_event_status}
\title{Create Event Status}
\usage{
set_event_status(data_set, outcome_date_col, censoring_date_col = NULL)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. NULL by default.}
}
\value{
Status
}
\description{
This function generates a binary status (1,0) associated
with an outcome. The value returned is 0 if a register is censored before
the outcome occurs or if the outcome does not occur during the follow-up
period. If the outcome occurs and the register is not censored, the value
returned is 1.
}
\keyword{internal}
