% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_effectiveness.R
\name{st_effectiveness}
\alias{st_effectiveness}
\title{Calculate vaccine effectiveness with the screening method}
\usage{
st_effectiveness(
  data,
  vaccinated_population,
  unvaccinated_population,
  vaccinated_outcome,
  unvaccinated_outcome,
  interval,
  group_by = NULL
)
}
\arguments{
\item{data}{A data frame with case, hospitalisation, and ICU information}

\item{vaccinated_population}{Character string with the name of the data column
with the vaccinated population data}

\item{unvaccinated_population}{Character string with the name of the data column
with the unvaccinated population data}

\item{vaccinated_outcome}{Character string with the name of the data column
with the outcome data for vaccinated population data}

\item{unvaccinated_outcome}{Character string with the name of the data column
with the outcome data for unvaccinated population data}

\item{interval}{Character string with the name of the data column
with the time interval (e.g. week)}

\item{group_by}{Character string with the name of the data column
with the variable to group by, or NULL}
}
\value{
A data frame with vaccine effectiveness over time and optionally for
each group
}
\description{
Calculate vaccine effectiveness with the screening method
}
\examples{
filtered_data <- readRDS(
  system.file(
    "extdata", "filtered_data.rds",
    package = "vaccineff",
    mustWork = TRUE
  )
)
st_effectiveness(
  data = filtered_data,
  vaccinated_population = "pob_comp",
  unvaccinated_population = "pob_no",
  vaccinated_outcome = "casos_comp",
  unvaccinated_outcome = "casos_no",
  interval = "fecha"
)
}
