% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{summary.effectiveness}
\alias{summary.effectiveness}
\title{Summarize VE Results}
\usage{
\method{summary}{effectiveness}(object, ...)
}
\arguments{
\item{object}{An object of class \code{effectiveness}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A summary of the results from effectiveness.
}
\description{
Summarizes the results of \code{effectiveness}.
}
