% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_vaccineff.R
\name{summary.vaccineff}
\alias{summary.vaccineff}
\title{Summarize VE Results}
\usage{
\method{summary}{vaccineff}(object, ...)
}
\arguments{
\item{object}{Object of the class \code{vaccineff}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Summary of the results from \code{estimate_vaccineff}.
}
\description{
Summarizes the results of \code{vaccineff}.
}
