% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vaccineff_data.R
\name{summary.vaccineff_data}
\alias{summary.vaccineff_data}
\title{Summarize Vaccineff Data}
\usage{
\method{summary}{vaccineff_data}(object, warnings_log = FALSE, ...)
}
\arguments{
\item{object}{Object of the class \code{vaccineff_data}.}

\item{warnings_log}{If \code{TRUE}, prints the warnings log.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Summary of the results from vaccineff data
}
\description{
Summarizes the results of \code{make_vaccineff_data}.
}
