% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnegdata.R
\docType{data}
\name{testnegdata}
\alias{testnegdata}
\title{Test-negative data on vaccineff}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
testnegdata
}
\description{
Data from
}
\examples{
testnegdata
}
\keyword{datasets}
