% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{truncate_from_start_cohort}
\alias{truncate_from_start_cohort}
\title{Internal function to truncate data based on start_cohort}
\usage{
truncate_from_start_cohort(
  data_set,
  outcome_date_col,
  censoring_date_col,
  start_cohort
)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. NULL by default.}
}
\value{
\code{data.frame} with truncated data
}
\description{
Internal function to truncate data based on start_cohort
}
\keyword{internal}
