% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{truncate_tte_at}
\alias{truncate_tte_at}
\title{Internal function to truncate time to events}
\usage{
truncate_tte_at(data_set, outcome_date_col, end_cohort, at)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information (see example).}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{end_cohort}{End date of the study.}

\item{at}{Time to truncate the follow-up period}
}
\value{
\code{data.frame} with truncated data
}
\description{
Internal function to truncate time to events
}
\keyword{internal}
