% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{adjust_time_to_event}
\alias{adjust_time_to_event}
\title{Internal function to truncate data based on start_cohort}
\usage{
adjust_time_to_event(data_set)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information (see example).}
}
\value{
\code{data.frame} with truncated data
}
\description{
Internal function to truncate data based on start_cohort
}
\keyword{internal}
