% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_utils.R
\name{balance_summary}
\alias{balance_summary}
\title{Balance of Vaccinated/Unvaccinated Groups}
\usage{
balance_summary(
  data,
  nearest,
  exact,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status
)
}
\arguments{
\item{data}{A data frame to assess matching balance.}

\item{nearest}{named vector with name(s) of column(s) for \code{nearest}
matching and caliper(s) for each variable.
e.g. \code{nearest = c("characteristic1" = n1, "characteristic2" = n2)},
where \code{n1} and \code{n2} are the calipers. Default to \code{NULL}.}

\item{exact}{name(s) of column(s) for \code{exact} matching.
Default to \code{NULL}.}

\item{vacc_status_col}{name of the column containing the vaccination.
Default is \code{vaccine_status}.}

\item{vaccinated_status}{string assigned to the vaccinated population.
Default is \code{v}.}

\item{unvaccinated_status}{string assigned to the unvaccinated population.
Default is \code{u}.}
}
\value{
A summary data frame with the balance of each variable by
vaccine status. Numeric variables are reported with means, and
categorical/factor variables are reported with proportions.
In both cases, the Standardized Mean Difference (SMD) is calculated.
}
\description{
This function creates a summary after matching.
}
\keyword{internal}
