% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{coh_coverage}
\alias{coh_coverage}
\title{Function to calculate the vaccine coverage per dose}
\usage{
coh_coverage(
  data,
  vacc_date_col,
  unit = c("day", "month", "year"),
  date_interval = FALSE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{vacc_date_col}{name of the column(s) that contains the vaccine date
to calculate the coverage}

\item{unit}{aggregation unit, must be either "year" or "month" or "day"}

\item{date_interval}{if FALSE, the function calculates the coverage interval
based on the min() and max() of the vacc_date_col.
It is also possible to pass a custom date interval to truncate or expand the
date interval (see example)}
}
\value{
data.frame with vaccine number of doses per date, cumulative count
of doses and vaccine coverage
}
\description{
This function returns the vaccination coverage of a dose along
the cohort study. The coverage can be calculated grouped by
year, day and month. This most by specified in the parameter unit.
To group by year, the dataset must contain more than one year of
historical data.
}
\examples{
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)
coh_coverage(
  data = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  date_interval = date_interval
)
}
