% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_hr.R
\name{coh_eff_hr}
\alias{coh_eff_hr}
\title{Function to Estimate the Vaccine Effectiveness Using Hazard Ratio}
\usage{
coh_eff_hr(
  data_set,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status,
  start_cohort,
  end_cohort
)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information (see example).}

\item{outcome_status_col}{Name of the column containing status of the event
(must be a binary column). Default is \code{outcome_status}.}

\item{time_to_event_col}{Name of the column containing the time-to-event.
Default is \code{time_to_event}.}

\item{vacc_status_col}{Name of the column containing the vaccination.
Default is \code{vaccine_status}.}

\item{vaccinated_status}{Status assigned to the vaccinated population.
Default is \code{v}.}

\item{unvaccinated_status}{Status assigned to the unvaccinated population.
Default is \code{u}.}

\item{start_cohort}{Start date of the study.}

\item{end_cohort}{End date of the study.}
}
\value{
Method call from \code{{survival}}, VE (CI95\%),
p-value of Schoenfeld test,
and log-log plot for proportional hazards
}
\description{
The function relies on the implementation of the Cox model for
proportional hazards from the package \code{{survival}} The proportional hazards
hypothesis is tested using the Schoenfeld test, and the resultant p-value is
provided in the results. Log-log plots are also calculated using the
Kaplan-Meier survival estimator to provide a visual test for the proportional
hazards hypothesis. The function returns a list with the method called, the
estimation of VE (CI95\%), the p-value of the Schoenfeld test, and the log-log
plot.
}
\keyword{internal}
