% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{coh_eff_noconf}
\alias{coh_eff_noconf}
\title{Function to estimate the vaccine effectiveness based on the vaccination
status.
The function relies on the implementation of the Cox model for proportional
hazards on survival package. It returns a dataframe with the summary of the
estimation that includes the value of the hazard ratio, the vaccine
effectiveness, and the result of the test for the Proportional Hazards
hypothesis. If the test is rejected, it is recomended to try other
statistical strategies, e.g. stratifying the dataset by confounders or
including time-dependent variables.}
\usage{
coh_eff_noconf(
  data,
  outcome_status_col,
  time_to_event_col,
  status_vacc_col,
  p_thr = 0.05
)
}
\arguments{
\item{data}{dataset with at least one column to generate the status}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{status_vacc_col}{name of the column containing the vaccination
status}

\item{p_thr}{p-value to test Proportional Hazards Hypothesis}
}
\value{
age_group
}
\description{
Function to estimate the vaccine effectiveness based on the vaccination
status.
The function relies on the implementation of the Cox model for proportional
hazards on survival package. It returns a dataframe with the summary of the
estimation that includes the value of the hazard ratio, the vaccine
effectiveness, and the result of the test for the Proportional Hazards
hypothesis. If the test is rejected, it is recomended to try other
statistical strategies, e.g. stratifying the dataset by confounders or
including time-dependent variables.
}
\examples{
\dontrun{
cohortdata <- data.frame()

cohortdata$immunization.death <- get_immunization_date(cohortdata,
"death.date", 0, 14,
c("vaccine.date.1", "vaccine.date.2"),
"2021-12-31", take_first = FALSE)
"2021-12-31",
take_first = FALSE)
cohortdata$vaccine.status <- set_status(cohortdata,
c("immunization.death"),
status = c("v", "u"))
cohortdata$death.status <- set_status(cohortdata,
c("death.date"))
cohortdata$time.to.death <- get_time_to_event(cohortdata, "death.date",
"2021-01-01", "2021-12-31",
FALSE)
coh_eff_noconf(cohortdata,
"death.status",
"time.to.death",
"vaccine.status")
}
}
