% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{coh_eff_noconf}
\alias{coh_eff_noconf}
\title{Function to estimate the vaccine effectiveness based on the vaccination
status.}
\usage{
coh_eff_noconf(
  data,
  outcome_status_col,
  time_to_event_col,
  status_vacc_col,
  p_thr = 0.05
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{status_vacc_col}{name of the column containing the vaccination
status}

\item{p_thr}{p-value to test Proportional Hazards Hypothesis}
}
\value{
summary: hazards ratio (CI95\%), vaccine effectiveness (CI95\%),
and Schoenfeld test
}
\description{
The function relies on the implementation of the Cox model for proportional
hazards on survival package. It returns a dataframe with the summary of the
estimation that includes the value of the hazard ratio (CI95\%), the vaccine
effectiveness (CI95\%), and the result of the test for the Proportional
Hazards hypothesis. If the test is rejected, it is recomended to try other
statistical strategies, e.g. stratifying the dataset by confounders or
including time-dependent variables.
}
\examples{
\dontrun{
data("cohortdata")
cohortdata$immunization_death <- get_immunization_date(
  date = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunisation_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = "2021-12-31",
  take_first = FALSE
)

cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("immunization_death"),
  status = c("v", "u")
)

cohortdata$death_status <- set_status(
  data = cohortdata,
  col_names = c("death_date")
)

cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = "2021-01-01",
  end_cohort = "2021-12-31",
  FALSE
)

coh_eff_noconf(
  cohortdata,
  "death_status",
  "time_to_death",
  "vaccine_status"
)
}
}
