% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortdata.R
\docType{data}
\name{cohortdata}
\alias{cohortdata}
\title{Cohort data on vaccineff}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
cohortdata
}
\description{
Data from
}
\examples{
cohortdata
}
\keyword{datasets}
