% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{cox_model}
\alias{cox_model}
\title{Internal function to calculate Cox-PH model and related metrics.}
\usage{
cox_model(
  data,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{vacc_status_col}{name of the column containing the vaccination}

\item{vaccinated_status}{string assigned to the vaccinated population}

\item{unvaccinated_status}{string assigned to the unvaccinated population}
}
\value{
List with data from Cox model:
hr - hazard ratio (CI95\%)
p_value
survival object with model
survival object with Schoenfeld test
}
\description{
Internal function to calculate Cox-PH model and related metrics.
}
\keyword{internal}
