% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{cox_model}
\alias{cox_model}
\title{Internal function to calculate Cox-PH model and related metrics.}
\usage{
cox_model(
  data_set,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status
)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information (see example).}

\item{outcome_status_col}{Name of the column containing status of the event
(must be a binary column). Default is \code{outcome_status}.}

\item{time_to_event_col}{Name of the column containing the time-to-event.
Default is \code{time_to_event}.}

\item{vacc_status_col}{Name of the column containing the vaccination.
Default is \code{vaccine_status}.}

\item{vaccinated_status}{Status assigned to the vaccinated population.
Default is \code{v}.}

\item{unvaccinated_status}{Status assigned to the unvaccinated population.
Default is \code{u}.}
}
\value{
List with data from Cox model:
hr - hazard ratio (CI95\%)
p_value
\code{{survival}} object with model
\code{{survival}} object with Schoenfeld test
}
\description{
Internal function to calculate Cox-PH model and related metrics.
}
\keyword{internal}
