% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match.R
\name{dataset.match}
\alias{dataset.match}
\title{Function for Extracting Matched Dataset}
\usage{
\method{dataset}{match}(object, ...)
}
\arguments{
\item{object}{An object of class \code{match}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
The dataset extracted from \code{match_cohort}.
}
\description{
This function extracts the matched datasets generated
by \code{match_cohort}.
}
