% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn_effectiveness.R
\name{estimate_tn_eff}
\alias{estimate_tn_eff}
\title{Estimate vaccine effectiveness using the test negative design}
\usage{
estimate_tn_eff(data, status_vacc_col, outcome_status_col)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{status_vacc_col}{name of the column containing the vaccination
status}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}
}
\value{
The vaccine effectiveness value
}
\description{
Estimate vaccine effectiveness using the test negative design
}
\details{
Currently, this function requires the arguments \code{status_vacc_col} and
\code{outcome_status_col} to be booleans (TRUE/FALSE). This will be relaxed
in the future to allow any kind of binary values for the two columns.
}
\examples{
# Load the dataset for this example
data(testnegdata)
testnegdata$pcr_result_bin <- set_binary_test(testnegdata,
                                              "pcr_result",
                                              "Neg",
                                              c(FALSE, TRUE))
testnegdata$vacc_status_bin <- set_binary_test(testnegdata,
                                              "vacc_status",
                                              "No",
                                              c(FALSE, TRUE))
estimate_tn_eff(testnegdata, "vacc_status_bin", "pcr_result_bin")
}
