% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{extract_surv_model}
\alias{extract_surv_model}
\title{Internal function to extract summary output from survival models.}
\usage{
extract_surv_model(model, start_cohort, end_cohort)
}
\arguments{
\item{model}{survival object with model}

\item{start_cohort}{start date of the study}

\item{end_cohort}{end date of the study
status}
}
\value{
Data frame with survival data
}
\description{
Internal function to extract summary output from survival models.
}
\keyword{internal}
