% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_adjust_expo.R
\name{get_exposition_time}
\alias{get_exposition_time}
\title{Calculate Exposition Time After Matching}
\usage{
get_exposition_time(data_set, outcome_date_col, censoring_date_col, end_cohort)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information.}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. \code{NULL} by default.}

\item{end_cohort}{End date of the study.}
}
\value{
\code{column} with exposition time per indivual.
}
\description{
This auxiliar function calculates the exposition time of
individuals after matching to ensure they coincide. Creating censoring
after match and defining t0_follow_up from pairing information is necessary
before using it.
}
\keyword{internal}
