% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_immunization_date}
\alias{get_immunization_date}
\title{Function to construct the inmunization date}
\usage{
get_immunization_date(
  data,
  outcome_date_col,
  outcome_delay,
  immunization_delay,
  vacc_date_col,
  end_cohort,
  take_first = TRUE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_date_col}{name of the column that contains
the outcome dates}

\item{outcome_delay}{characteristic time in days of the outcome
from the reference event}

\item{immunization_delay}{characteristic time in days before the patient
is considered immune}

\item{vacc_date_col}{name of the column(s) that contains the vaccine dates}

\item{end_cohort}{end date of the study}

\item{take_first}{TRUE: takes the minimum vaccine date for
registers without outcome.
FALSE: takes closest to end_cohort}
}
\value{
immunzation date
}
\description{
This function returns a column with the immunization date that corresponds
to an analysed outcome.
If a register presents an outcome, the function search for the closest
vaccine date before the outcome that satisfies the condition:
vacc_date_col <= outcome_date_col - delay_time - immunization_delay.
This condition allows to discriminate the vaccine dates in
terms of characteristic times in days (delay_time)  associated to an
outcome, from the onset of symptoms or from any reference event, and the
characteristic time in days before the patient is considered
immune (immunization_delay).
Both parameters can be set to zero by the user without
affecting the results.
If a register does not present an outcome, the immunization
date can be construct using the closest vaccine
date to the end of the study (take_first = FALSE), or the
first vaccination date found (take_first = TRUE).
Notice that the function works for one or several vaccines.
In case of several vaccines, the parameter
{outcome_date_col must} be passed as a vector (see example)
}
\examples{
\dontrun{
data("cohortdata")
cohortdata$immunization.death <- get_immunization_date(cohortdata,
"death_date", 1, 1,
c("vaccine_date_1", "vaccine_date_2"),
"2021-12-31",
take_first = FALSE)
}
}
