% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_immunization_dose}
\alias{get_immunization_dose}
\title{Function to construct dose associated to the immunization date}
\usage{
get_immunization_dose(
  data,
  immunization_date_col,
  vacc_date_col,
  immunization_delay
)
}
\arguments{
\item{data}{dataset with at least one outcome column to generate the time-to-event}

\item{immunization_date_col}{DESCRIBE!}

\item{vacc_date_col}{DESCRIBE!}

\item{immunization_delay}{DESCRIBE!}
}
\value{
dose
}
\description{
This function returns the names of the columns associated to the immunization date.
To avoid mistakes, it is necessary to set the same value of immunization_delay that was used in the previous
functions.
}
\examples{
\dontrun{
data("cohortdata")
cohortdata$immunization.death <- get_immunization_date(cohortdata, "death.date", 1, 1,
                                                         c("vaccine.date.1", "vaccine.date.2"), 
                                                         "2021-12-31", take_first = FALSE)
cohortdata$immunization.dose <- get_immunization_dose(cohortdata, "immunization.death", 
                                                        c("vaccine.date.1", "vaccine.date.2"),
                                                        immunization_delay = 14)
}
}
