% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_immunization_dose}
\alias{get_immunization_dose}
\title{Function to construct dose associated to the immunization date}
\usage{
get_immunization_dose(
  data,
  immunization_date_col,
  vacc_date_col,
  immunization_delay
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{immunization_date_col}{name of the column that contains the
immunization date.}

\item{vacc_date_col}{name of the column(s) that contains the vaccine date}

\item{immunization_delay}{characteristic time in days before the patient
is considered immune}
}
\value{
dose: a column with the names of the columns that are associated to
the doses of each register
}
\description{
This function returns the names of the columns associated
to the immunization date.
To avoid mistakes, it is necessary to set the same value of
immunization_delay that was used in the previous functions.
}
\examples{
\dontrun{
data("cohortdata")
cohortdata$immunization_death <- get_immunization_date(
  data = cohortdata,
  outcome_date_col = "death_date",
  outcome_delay = 0,
  immunization_delay = 14,
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  end_cohort = as.Date("2021-12-31"),
  take_first = FALSE
)
cohortdata$immunization_dose <- get_immunization_dose(
  data = cohortdata,
  immunization_date_col = "immunization_death",
  vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
  immunization_delay = 14
)
}
}
