% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_immunization_vaccine}
\alias{get_immunization_vaccine}
\title{Function to construct vaccine biologic associated to the immunization date}
\usage{
get_immunization_vaccine(
  data,
  immunization_date_col,
  vacc_date_col,
  vacc_name_col,
  immunization_delay
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{immunization_date_col}{name of the column that contains the
immunization date.}

\item{vacc_date_col}{name of the column(s) that contains the vaccine date}

\item{vacc_name_col}{name of the column(s) that contains the vaccine names}

\item{immunization_delay}{characteristic time in days before the patient
is considered immune}
}
\value{
dose: a column with the names of the columns that are associated to
the doses of each register

vaccine name
}
\description{
This function returns the names of the vaccines
associated to the immunization date.
To avoid mistakes, it is necessary to set the same
value of immunization_delay that was used in the previous
functions.
The arguments vacc_date_col and vacc_name_col must be passed in the same
order, i.e. every name of column date must correspond
to a name of vaccine column (see example)
}
\examples{
\dontrun{
data("cohortdata")
cohortdata$immunization_death
  <- get_immunization_date(cohortdata,
                           "death_date",
                           1,
                           1,
                           c("vaccine_date_1","vaccine_date_2"),
                           "2021-12-31",
                           take_first = FALSE)
cohortdata$immunization.vaccine
  <- get_immunization_vaccine(cohortdata,
                              "immunization_death",
                              c("vaccine_date_1", "vaccine_date_2"),
                              c("vaccine_1", "vaccine_2"),
                              immunization_delay = 14)
}
}
