% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_immunization_vaccine}
\alias{get_immunization_vaccine}
\title{Function to construct vaccine biologic associated to the immunization date}
\usage{
get_immunization_vaccine(
  data,
  immunization_date_col,
  vacc_date_col,
  vacc_name_col,
  immunization_delay
)
}
\arguments{
\item{data}{dataset with at least one outcome column to
generate the time-to-event}

\item{immunization_date_col}{DESCRIBE!}

\item{vacc_date_col}{DESCRIBE!}

\item{vacc_name_col}{DESCRIBE!}

\item{immunization_delay}{DESCRIBE!}
}
\value{
vaccine type
}
\description{
This function returns the names of the vaccines
associated to the immunization date.
To avoid mistakes, it is necessary to set the same
value of immunization_delay that was used in the previous
functions.
The arguments vacc_date_col and vacc_name_col in the same
order, i.e. every name of column date must correspond
to a name of vaccine column (see example)
}
\examples{
\dontrun{
cohortdata <- data.frame()
cohortdata$immunization.death
  <- get_immunization_date(cohortdata,
                           "death.date",
                           1,
                           1,
                           c("vaccine.date.1","vaccine.date.2"),
                           "2021-12-31",
                           take_first = FALSE)
cohortdata$immunization.vaccine
  <- get_immunization_vaccine(cohortdata,
                              "immunization.death",
                              c("vaccine.date.1", "vaccine.date.2"),
                              c("vaccine.1", "vaccine.2"),
                              immunization_delay = 14)
}
}
