% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{get_time_to_event}
\alias{get_time_to_event}
\title{Construct Time-to-Event}
\usage{
get_time_to_event(
  data,
  outcome_date_col,
  censoring_date_col = NULL,
  start_cohort,
  end_cohort,
  start_from_immunization = FALSE,
  immunization_date_col
)
}
\arguments{
\item{data}{Dataset with cohort information (see example).}

\item{outcome_date_col}{Name of the column that contains the outcome dates.}

\item{censoring_date_col}{Name of the column that contains the censoring
date. NULL by default.}

\item{start_cohort}{Start date of the study.}

\item{end_cohort}{End date of the study.}

\item{start_from_immunization}{TRUE: starts counting time-to-event from the
immunization date if available. FALSE: starts counting time-to-event from
the start date of the cohort study.}

\item{immunization_date_col}{Name of the column that contains the
immunization date. Required if start_from_immunization = TRUE.}
}
\value{
time-to-event
}
\description{
This function returns a column with the time-to-event in days
until a reference outcome. The starting point to count the time-to-event
can be the immunization date, assuming that the vaccinated population enters
the study when they are vaccinated (start_from_immunization=TRUE). Or the
beginning of the study, if all the cohort is known at this point
(start_from_immunization=FALSE). In this last case, it is not necessary to
pass the argument \code{immunization_date_col}.
}
\keyword{internal}
