% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{km_model}
\alias{km_model}
\title{Internal function to calculate Kapplan-Meier model and related metrics.}
\usage{
km_model(
  data,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status,
  start_cohort,
  end_cohort
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{vacc_status_col}{name of the column containing the vaccination}

\item{vaccinated_status}{string assigned to the vaccinated population}

\item{unvaccinated_status}{string assigned to the unvaccinated population}

\item{start_cohort}{start date of the study}

\item{end_cohort}{end date of the study
status}
}
\value{
Data frame with data from KM model:
time to event
survivial probability (CI95\%)
cumulative incidence (CI95\%)
}
\description{
Internal function to calculate Kapplan-Meier model and related metrics.
}
\keyword{internal}
