% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_utils.R
\name{km_model}
\alias{km_model}
\title{Internal function to calculate Kaplan-Meier model and related metrics.}
\usage{
km_model(
  data_set,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status,
  unvaccinated_status,
  start_cohort,
  end_cohort
)
}
\value{
\code{data.frame} with data from KM model:
"time", "date", "strata",
"n.risk", "n.event", "n.censor",
"surv", "lower", "upper",
"cumincidence", "cumincidence_lower", "cumincidence_upper"
}
\description{
Internal function to calculate Kaplan-Meier model and related metrics.
}
\keyword{internal}
