% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_matching.R
\name{match_cohort}
\alias{match_cohort}
\title{Static match of cohort}
\usage{
match_cohort(
  data,
  status_vacc_col,
  method = c("exact", "nearest"),
  exact = NULL,
  nearest = NULL,
  caliper = NULL
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{status_vacc_col}{name of the column containing the information
of the vaccination status.}

\item{method}{matching method. \code{exact} and \code{nearest} are allowed.}

\item{exact}{name(s) of column(s) for \code{exact} matching.
Default to \code{NULL}.}

\item{nearest}{name(s) of column(s) for \code{nearest} matching.
Default to \code{NULL}.}

\item{caliper}{named vector with caliper(s).
e.g. \verb{nearest = c("characteristic1", "characteristic2"), caliper = c("characteristic1" = n1, "characteristic2" = n2)},
where \code{n1} and \code{n2} are the calipers.}
}
\value{
data frame with matched population. Two columns are added
to the structure provided in \code{data}:
\code{prop_score} (propensity score of the match),
\code{subclass} (id of matched couple)
}
\description{
This function builds couples of vaccinated - unvaccinated
individuals with similar characteristics. The function relies on  the
propensity score matching algorithm implemented in MatchIt package.
Thus, it allows for \code{exact} and \code{nearest} methods.
When \code{method = "exact"} the function returns couples with the same
characteristics for the variables passed in the parameter \code{exact}.
In this case, any inputs in the parameter \code{nearest} are ignored.
The \code{nearest} method can be used both with \code{nearest} and exact coincidences.
The exact characteristics must be provided as inputs in the
parameter \code{exact} and the others as inputs in \code{nearest}. In addition,
the calipers must be provided as a named vector for each of the variables
provided in \code{nearest}
(e.g. \verb{nearest = c("characteristic1", "characteristic2"), caliper = c("characteristic1" = n1, "characteristic2" = n2)},
where \code{n1} and \code{n2} are the calipers).
If no input is provided to \code{exact} when using \code{method = nearest }, only
near matches are returned.
}
\examples{
# load package example data for cohort studies
data("cohortdata")

# assign vaccination status
cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = c("vaccine_date_1", "vaccine_date_2"),
  status = c("v", "u")
)

# match cohort
cohort_match <- match_cohort(data = cohortdata,
  status_vacc_col = "vaccine_status",
  method = "nearest",
  nearest = "age",
  exact = "sex",
  caliper = c("age" = 1)
)

# view data with added columns
head(cohortdata)
}
