% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match_utils.R
\name{match_cohort}
\alias{match_cohort}
\title{Match cohort using mahalanobis distance}
\usage{
match_cohort(data, vacc_status_col, exact = NULL, nearest = NULL)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{vacc_status_col}{name of the column containing the vaccination}

\item{exact}{name(s) of column(s) for \code{exact} matching.
Default to \code{NULL}.}

\item{nearest}{named vector with name(s) of column(s) for \code{nearest}
matching and caliper(s) for each variable.
e.g. \code{nearest = c("characteristic1" = n1, "characteristic2" = n2)},
where \code{n1} and \code{n2} are the calipers. Default to \code{NULL}.}
}
\value{
data frame with matched population. Two columns are added
to the structure provided in \code{data}:
\code{prop_score} (propensity score of the match),
\code{subclass} (id of matched couple)
}
\description{
This function builds couples of vaccinated - unvaccinated
individuals with similar characteristics. The function relies on  the
matching algorithm implemented in the package \code{{MatchIt}}.
By default the function uses \code{method = "nearest"}, \code{ratio = 1},
\code{distance = "mahalanobis"} to match the data.
Exact and near characteristics are accepted for the matching criteria.
These are passed in the parameters \code{exact} and \code{nearest}, respectively.
The parameter \code{nearest} must be provided together with the calipers
following as a named vector.
(e.g. \verb{nearest = c("characteristic1", "characteristic2"), caliper = c(characteristic1 = n1, characteristic2 = n2)},
where \code{n1} and \code{n2} are the calipers).
}
\keyword{internal}
