% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_effectiveness.R
\name{plot.effectiveness}
\alias{plot.effectiveness}
\title{Function for Extracting VE plot}
\usage{
\method{plot}{effectiveness}(effectiveness)
}
\arguments{
\item{effectiveness}{An object of class \code{effectiveness}.}
}
\value{
Plot extracted from \code{effectiveness}.
}
\description{
This function extracts the plot generated
by \code{effectiveness}.
}
