% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_covegare.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot Vaccine Coverage}
\usage{
plot_coverage(
  data_set,
  vacc_date_col,
  unit = c("day", "month", "year"),
  doses_count_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = NULL,
  cumulative = FALSE
)
}
\arguments{
\item{data_set}{\code{data.frame} with cohort information (see example).}

\item{vacc_date_col}{Name of the column(s) that contain the vaccine date to
calculate the coverage.}

\item{unit}{Aggregation unit, must be either "year", "month", or "day".}

\item{doses_count_color}{Color assigned to the doses count.}

\item{coverage_color}{Color assigned to the coverage calculation.}

\item{date_interval}{If NULL, the function calculates the coverage interval
based on the min() and max() of the \code{vacc_date_col}. It is also possible to
pass a custom date interval to truncate or expand the date interval (see
example).}

\item{cumulative}{If \code{TRUE}, returns the cumulative number of doses over the
time window.}
}
\value{
2-axis ggplot2 plot of vaccine coverage and daily doses.
}
\description{
This function returns a plot of the vaccine coverage or the
cumulative coverage (if cumulative = TRUE). The return is a 2-axis \code{ggplot2}
element with the number of vaccines per date on the left axis and the
coverage per date on the right axis. Colors can be defined with the
parameters of the function \code{doses_count_color} and \code{coverage_color}.
Other graphic elements such as labels and legends can be manipulated
using \code{ggplot2}.
}
\examples{
# Load example data
data("cohortdata")

# Define date intervals for coverage
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)

# Plot coverage
plot_coverage(
  data_set = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  doses_count_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = date_interval,
  cumulative = FALSE
)
}
