% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Function to plot the vaccine coverage in a cohort}
\usage{
plot_coverage(
  data,
  vacc_date_col,
  unit = c("day", "month", "year"),
  doses_count_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = NULL,
  cumulative = FALSE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{vacc_date_col}{name of the column(s) that contains the vaccine date
to calculate the coverage}

\item{unit}{aggregation unit, must be either "year" or "month" or "day"}

\item{doses_count_color}{color assigned to the doses count}

\item{coverage_color}{color assigned to the coverage calculation}

\item{date_interval}{if NULL, the function calculates the coverage interval
based on the min() and max() of the vacc_date_col.
It is also possible to pass a custom date interval to truncate or expand the
date interval (see example)}

\item{cumulative}{if TRUE returns the cumulative number of doses
over the time window}
}
\value{
2-axis ggplot2 plot of vaccine coverage and daily doses
}
\description{
The function returns a plot of the vaccine coverage or the cumulative
coverage (if cumulative = TRUE).
The return is a 2-axis ggplot2 element with the number of vaccines per date
in the left-axes and the coverage per date in the right-axes.
It is possible to manipulate the colors, labels, legend and most of
the graphic elements.
}
\examples{
data("cohortdata")
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")
date_interval <- c(start_cohort, end_cohort)
plot_coverage(
  data = cohortdata,
  vacc_date_col = "vaccine_date_1",
  unit = "month",
  doses_count_color = "steelblue",
  coverage_color = "mediumpurple",
  date_interval = date_interval,
  cumulative = FALSE
)
}
