% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_plot.R
\name{plot_loglog}
\alias{plot_loglog}
\title{Plot Log-Log Test for Proportional Hazards Hypothesis}
\usage{
plot_loglog(cox_model_prediction)
}
\arguments{
\item{cox_model_prediction}{Prediction from Cox model.}
}
\value{
Log-log plot.
}
\description{
This function uses the return from the Cox model
to create a log-log plot.
}
\keyword{internal}
