% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_plot.R
\name{plot_loglog}
\alias{plot_loglog}
\title{Plot Log-Log Test for Proportional Hazards Hypothesis}
\usage{
plot_loglog(km)
}
\arguments{
\item{km}{Kaplan-Meier estimation created with \code{km_model}.}
}
\value{
Log-log plot.
}
\description{
This function uses the return from the Kaplan-Meier model
to create a log-log plot. It calculates log(-log(Survival Probability))
and log(Time).
}
\keyword{internal}
