% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_eff_plot.R
\name{plot_loglog}
\alias{plot_loglog}
\title{Plot Log-Log Test for Proportional Hazards Hypothesis}
\usage{
plot_loglog(km, vaccinated_status, unvaccinated_status)
}
\arguments{
\item{vaccinated_status}{string assigned to the vaccinated population.
Default is \code{v}.}

\item{unvaccinated_status}{string assigned to the unvaccinated population.
Default is \code{u}.}

\item{surv_plot}{Survival plot generated by \code{plot_survival}.}
}
\value{
Log-log plot.
}
\description{
This function uses the return from the Kaplan-Meier model
to create a log-log plot. It calculates log(-log(Survival Probability))
and log(Time).
}
\keyword{internal}
