% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_plot.R
\name{plot_survival}
\alias{plot_survival}
\title{Function to plot the Survival probability based on the Kaplan-Meir model}
\usage{
plot_survival(
  data,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  vaccinated_status = "v",
  unvaccinated_status = "u",
  vaccinated_color = "steelblue",
  unvaccinated_color = "darkred",
  start_cohort,
  end_cohort,
  percentage = TRUE,
  cumulative = FALSE,
  loglog = FALSE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{vacc_status_col}{name of the column containing the vaccination
status}

\item{vaccinated_status}{string assigned to the vaccinated population}

\item{unvaccinated_status}{string assigned to the unvaccinated population}

\item{vaccinated_color}{color assigned to the vaccinated population}

\item{unvaccinated_color}{color assigned to the unvaccinated population}

\item{start_cohort}{start date of the study}

\item{end_cohort}{end date of the study}

\item{percentage}{if TRUE returns probability in percentage}

\item{cumulative}{if TRUE returns cumulative Hazards}

\item{loglog}{if TRUE returns loglog plot for testing proportional hazards}
}
\value{
ggplot2 plot of survival/cumulative hazard
}
\description{
The function relies on the implementation of the Kaplan-Meier model from
survival package.
It returns a plot of the Survival Probability or the Cumulative Hazard (if
cumulative = TRUE).
The return is a ggplot2 element of the curves with 95\% C.I.
It is possible to manipulate the colors, labels, legend and most of
the graphic elements.
}
\examples{
# load example package data
data("cohortdata")
head(cohortdata)
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")

cohortdata$immunization <-
  get_immunization_date(
    data = cohortdata,
    outcome_date_col = "death_date",
    outcome_delay = 0,
    immunization_delay = 14,
    vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
    end_cohort = end_cohort,
    take_first = FALSE
  )

cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = "immunization",
  status = c("v", "u")
)

cohortdata$death_status <- set_status(
  data = cohortdata,
  col_names = "death_date"
)

cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  start_from_immunization = FALSE
)

plot_survival(
  data = cohortdata,
  outcome_status_col = "death_status",
  time_to_event_col = "time_to_death",
  vacc_status_col = "vaccine_status",
  vaccinated_status = "v",
  unvaccinated_status = "u",
  vaccinated_color = "steelblue",
  unvaccinated_color = "darkred",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  percentage = TRUE,
  cumulative = TRUE,
  loglog = FALSE
)
}
