% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_plot.R
\name{plot_survival}
\alias{plot_survival}
\title{Function to plot the Survival probability based on the Kaplan-Meir model.}
\usage{
plot_survival(
  data,
  outcome_status_col,
  time_to_event_col,
  vacc_status_col,
  status,
  start_cohort,
  end_cohort,
  colors = c(c1 = "steelblue", c2 = "darkred"),
  percentage = TRUE,
  cumulative = FALSE
)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_status_col}{name of the column containing status of the
event (most be a binary column)}

\item{time_to_event_col}{name of the column containing the time-to-event}

\item{vacc_status_col}{name of the column containing the vaccination
status}

\item{start_cohort}{start date of the study}

\item{end_cohort}{end date of the study}

\item{colors}{list of two colors the type:
c("c1" = "steelblue", "c2" = "darkred")}

\item{percentage}{if TRUE returns probability in percentage}

\item{cumulative}{if TRUE returns cumulative Hazards (1-Survival)}
}
\value{
Survival/Cumulative hazard plot: ggplot2 element
}
\description{
The function relies on the implementation of the Kaplan-Meier model from
survival package.
It returns a plot of the Survival Probability or the Cumulative Hazard (if
cumulative = TRUE).
The return is a ggplot2 element of the curves with 95\% C.I that can be
maniulated in order to change the colors, labels, legen, etc.
}
\examples{
# load example package data
data("cohortdata")
head(cohortdata)
start_cohort <- as.Date("2044-01-01")
end_cohort <- as.Date("2044-12-31")

cohortdata$immunization_death <-
  get_immunization_date(
    data = cohortdata,
    outcome_date_col = "death_date",
    outcome_delay = 0,
    immunization_delay = 14,
    vacc_date_col = c("vaccine_date_1", "vaccine_date_2"),
    end_cohort = end_cohort,
    take_first = FALSE
  )

cohortdata$vaccine_status <- set_status(
  data = cohortdata,
  col_names = "immunization_death",
  status = c("v", "u")
)

cohortdata$death_status <- set_status(
  data = cohortdata,
  col_names = "death_date"
)

cohortdata$time_to_death <- get_time_to_event(
  data = cohortdata,
  outcome_date_col = "death_date",
  start_cohort = start_cohort,
  end_cohort = end_cohort,
  start_from_immunization = FALSE
)

survival_plot <- plot_survival(data = cohortdata,
                               outcome_status_col = "death_status",
                               time_to_event_col = "time_to_death",
                               vacc_status_col = "vaccine_status",
                               status = c("v", "u"),
                               start_cohort = start_cohort,
                               end_cohort = end_cohort,
                               colors = c("steelblue", "darkred"),
                               percentage = TRUE,
                               cumulative = TRUE)
}
