% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_data_wrangling.R
\name{set_event_status}
\alias{set_event_status}
\title{Assign event status}
\usage{
set_event_status(data, outcome_date_col, censoring_date_col = NULL)
}
\arguments{
\item{data}{dataset with cohort information (see example)}

\item{outcome_date_col}{name of the column that contains
the outcome dates}

\item{censoring_date_col}{name of the column that contains
the censoring date. NULL by default}
}
\value{
status
}
\description{
This function generates a binary status (1,0) associated
with an outcome. The value returned is 0 if a register is censored
before the outcome occurs or if the outcome does not occur during the
follow-up period. If the outcome occurs and the register is not
censored the value returned is 1.
}
\examples{
# load package example data for cohort studies
data("cohortdata")

# assign death status
cohortdata$death_status <- set_event_status(
  data = cohortdata,
  outcome_date_col = "death_date",
  censoring_date_col = "death_other_causes"
)

# view data with added column
head(cohortdata)
}
