% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match.R
\name{summary.match}
\alias{summary.match}
\title{Summarize Matching Results}
\usage{
\method{summary}{match}(object, ...)
}
\arguments{
\item{object}{An object of class \code{match}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A summary of the results from matching.
}
\description{
Summarizes the results of \code{match_cohort}.
}
