% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coh_match.R
\name{summary.match}
\alias{summary.match}
\title{Summarize Matching Results}
\usage{
\method{summary}{match}(object, warnings_log = FALSE, ...)
}
\arguments{
\item{object}{Object of the class \code{match}.}

\item{warnings_log}{If \code{TRUE}, prints the warnings log.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Summary of the results from matching.
}
\description{
Summarizes the results of \code{match_cohort}.
}
