% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccineff_package.R
\docType{package}
\name{vaccineff-package}
\alias{vaccineff}
\alias{vaccineff-package}
\title{vaccineff: Estimation of the Vaccines Effectiveness Based on Different Designs}
\description{
A package to estimate the vaccines effectiveness based on different designs. following format: Author et al. (2023) \doi{10.5281/zenodo.6619350}.
}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co}

Authors:
\itemize{
  \item David Santiago Quevedo \email{ex-dsquevedo@javeriana.edu.co}
  \item Santiago Loaiza \email{santiago.loaiza@javeriana.edu.co}
}

Other contributors:
\itemize{
  \item Geraldine Gómez Millán \email{geralidine.gomez@javeriana.edu.co} [contributor]
}

}
\keyword{internal}
