% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccineff_package.R
\docType{package}
\name{vaccineff-package}
\alias{vaccineff}
\alias{vaccineff-package}
\title{vaccineff: Estimate Vaccine Effectiveness Based on Different Study Designs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for estimating vaccine effectiveness and related metrics. The 'vaccineff_data' class manages key features for preparing, visualizing, and organizing cohort data, as well as estimating vaccine effectiveness. The results and model performance are assessed using the 'vaccineff' class.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/vaccineff}
  \item \url{https://epiverse-trace.github.io/vaccineff/}
  \item Report bugs at \url{https://github.com/epiverse-trace/vaccineff/issues}
}

}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{tracelacdev@gmail.com} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID})

Authors:
\itemize{
  \item David Santiago Quevedo \email{ex-dsquevedo@javeriana.edu.co} (\href{https://orcid.org/0000-0003-1583-4262}{ORCID})
}

Other contributors:
\itemize{
  \item International Development Research Center (IDRC) (0445x0472) [funder]
  \item Geraldine Gómez Millán \email{geralidine.gomez@javeriana.edu.co} (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [contributor]
  \item Érika J Cantor \email{erika.cantor@javeriana.edu.co} (\href{https://orcid.org/0000-0003-3320-6032}{ORCID}) [contributor]
  \item Santiago Loaiza \email{santiago.loaiza@javeriana.edu.co} (\href{https://orcid.org/0000-0002-2092-3262}{ORCID}) [contributor]
  \item Jaime Pavlich-Mariscal \email{jpavlich@javeriana.edu.co} (\href{https://orcid.org/0000-0002-3892-6680}{ORCID}) [contributor]
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [contributor]
  \item Felipe Segundo Abril \email{fsabrilb@unal.edu.co} (\href{https://orcid.org/0000-0002-2512-4929}{ORCID}) [contributor]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [contributor]
  \item Julian Otero \email{jd.otero10@uniandes.edu.co} (\href{https://orcid.org/0009-0006-0429-7747}{ORCID}) [contributor]
  \item Pontificia Universidad Javeriana (03etyjw28) [copyright holder]
}

}
\keyword{internal}
