% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccineff_package.R
\docType{package}
\name{vaccineff-package}
\alias{vaccineff}
\alias{vaccineff-package}
\title{vaccineff: Estimate Vaccine Effectiveness Based on Different Study Designs}
\description{
Estimate vaccine effectiveness based on different observational study designs, as discussed in Torvaldsen and McIntyre (2020) \doi{10.3316/informit.511798489353134}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/vaccineff}
  \item \url{https://epiverse-trace.github.io/vaccineff/}
  \item Report bugs at \url{https://github.com/epiverse-trace/vaccineff/issues}
}

}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID})

Authors:
\itemize{
  \item David Santiago Quevedo \email{ex-dsquevedo@javeriana.edu.co} (\href{https://orcid.org/0000-0003-1583-4262)}{ORCID})
  \item Santiago Loaiza \email{santiago.loaiza@javeriana.edu.co} (\href{https://orcid.org/0000-0002-2092-3262}{ORCID})
}

Other contributors:
\itemize{
  \item Geraldine Gómez Millán \email{geralidine.gomez@javeriana.edu.co} [contributor]
}

}
\keyword{internal}
